extract/web/requirements.txt
python-dotenv
trafilatura
newspaper4k
ollama
requests
langchain-openai
langchain-core
